octeontx: domain reset on any device closure
A change in DTS was made, enabling new generic driver for SMMU,
which by default creates unique translation context for each PCI
device, whereas, previously, IOVA translation was bypassed.
If an ODP application exits unexpectedly, the order of device closure
is random, which may lead to releasing translation contexts before
performing domain reset.
So far domain reset has been triggered when the first SSO in domain
gets closed. It has been agreed we may now need to trigger reset on any
device closure (ODP side will keep all of them opened, so that there
will not be any random reset).

Reviewed-on: https://sj1git1.cavium.com/9373
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
[RH: Original patch taken from marvell 88F3720 board support SDK 10.0-PR2003]

+++ drivers/net/ethernet/cavium/octeontx-83xx/cpt.h
@@ -14,7 +14,7 @@
 #include "octeontx.h"
 
 struct cptpf_com_s {
-	int (*create_domain)(u32, u16, u32, struct kobject *kobj);
+	u64 (*create_domain)(u32, u16, u32, struct kobject *kobj);
 	int (*destroy_domain)(u32 id, u16 domain_id, struct kobject *kobj);
 	int (*reset_domain)(u32, u16);
 };
