+++ drivers/net/ethernet/cavium/octeontx-83xx/sso.h
@@ -284,7 +284,7 @@ struct ssowpf {
 };
 
 struct ssowpf_com_s {
-	int (*create_domain)(u32, u16, u32, void *, void *,
+	u64 (*create_domain)(u32, u16, u32, void *, void *,
 			     struct kobject *kobj);
 	int (*destroy_domain)(u32, u16, struct kobject *kobj);
 	int (*reset_domain)(u32, u16, u64);
