+++ drivers/net/ethernet/cavium/octeontx-83xx/tim.h
@@ -14,7 +14,7 @@
 #include "octeontx.h"
 
 struct timpf_com_s {
-	int (*create_domain)(u32 id, u16 domain_id, u32 num_vfs,
+	u64 (*create_domain)(u32 id, u16 domain_id, u32 num_vfs,
 			     struct octeontx_master_com_t *com, void *domain,
 			     struct kobject *kobj);
 	int (*destroy_domain)(u32 id, u16 domain_id, struct kobject *kobj);
